Move ETH SwapIdPrefix into serialization
